[Fix] 기존의 FindWithImage 컴포넌트를 분리 및 컴포넌트 명 변경
기존의 FindWithImage 컴포넌트에서 페이지와 모달을 분리하기 위해 SearchWithImage컴포넌트와 CaptureGuideModal컴포넌트로 분리
