+++ front/components/search/SearchWithImage.style.ts
