+++ front/pages/search/index.tsx
@@ -1,15 +1,19 @@
 import type { NextPage } from "next";
 import { GetServerSideProps } from "next";
 import { CAPTURE_GUIDE_MOAL_COOKIE } from "@utils/constant";
-import FindWithImage, { FindWithImageProp } from "@searchComp/FindWithImage";
+import Search from "@searchComp/index";
 import { NextSeo } from "next-seo";
 import { SEARCH_IMAGE_SEO } from "next-seo.config";
 
-const Image: NextPage<FindWithImageProp> = ({ foundCookie }) => {
+export interface SearchProp {
+  foundCookie: boolean;
+}
+
+const Image: NextPage<SearchProp> = ({ foundCookie }) => {
   return (
     <>
       <NextSeo {...SEARCH_IMAGE_SEO} />
-      <FindWithImage foundCookie={foundCookie} />
+      <Search foundCookie={foundCookie} />
     </>
   );
 };
